arithmos: created a gimbal-system library and matrix rotation
