+++ sdk/core/src/libs/electrostatic-primer/electronetsoft/arithmos/vector3d/vec3d_rotate.c
@@ -1,130 +1,7 @@
 #include <electronetsoft/arithmos/vectorspaces/vector2d/vector2d.h>
 #include <electronetsoft/arithmos/vectorspaces/vector3d/vector3d.h>
+#include <electronetsoft/arithmos/vectorspaces/gimbal_system.h>
 #include <stdio.h>
-
-struct rotation_metadata {
-    matrix *in_orientation;
-    matrix *out_orientation;
-};
-
-static inline void preprocess_orientator(vector3d *v, vector3d *axis) {
-
-    // create a column vector matrix
-    vec_component axis_x[1] = {0,};
-    vec_component axis_y[1] = {0,};
-    vec_component axis_z[1] = {0,};
-    vec_component *axis_comps[3] = {axis_x, axis_y, axis_z};
-    matrix __axis = {
-            .element = axis_comps,
-            .m = 3,
-            .n = 1
-    };
-
-    mat_processors procs = {
-    };
-    mat_create_from_vector3d(*axis, &__axis, procs);
-
-    // find the position of the orientation vectors
-    vec_component _axis_x[1] = {0,};
-    vec_component _axis_y[1] = {0,};
-    vec_component _axis_z[1] = {0,};
-    vec_component *_axis_comps[3] = {_axis_x, _axis_y, _axis_z};
-    matrix ___axis = {
-            .element = _axis_comps,
-            .m = 3,
-            .n = 1
-    };
-    mat_product(*(v->gimbal.orientation), __axis, &___axis, procs);
-
-    axis->x = *(___axis.element[0]);
-    axis->y = *(___axis.element[1]);
-    axis->z = *(___axis.element[2]);
-}
-
-static inline status_code __on_entry_iterated(mat_proc_sig proc_sig) {
-    struct rotation_metadata *metadata = proc_sig.metadata;
-    matrix *in_orientation = metadata->in_orientation;
-    matrix *out_orientation = metadata->out_orientation;
-    uint64_t row_index = proc_sig.row_index;
-    uint64_t col_index = proc_sig.col_index;
-
-    out_orientation->element[row_index][col_index] =
-            in_orientation->element[row_index][col_index];
-
-    return PASS;
-}
-
-static inline status_code init_rotator_gimbal(vector3d axis,
-                                              matrix *__rotator,
-                                              vec_component angle,
-                                              vec_component *angle1,
-                                              vec3d_gimbal *gimbal) {
-
-    if (NULL == __rotator || NULL == __rotator->element) {
-        return EUNDEFINEDBUFFER;
-    }
-
-    if (get_vec_gimbal(axis) == GIMBAL_Z) {
-        // rotate around z-axis
-        // pre-processing automata
-        // Let the X-axis in the R(3) space be the X-axis in the 2D projection.
-        // Let the Y-axis in the R(3) space be the Y-axis in the 2D projection.
-
-        // work in XY plane
-        __rotator->element[0][0] = vector2d_cos(angle);
-        __rotator->element[0][1] = -vector2d_sin(angle);
-        __rotator->element[1][0] = vector2d_sin(angle);
-        __rotator->element[1][1] = vector2d_cos(angle);
-        __rotator->element[2][2] = 1;
-
-        if (NULL != gimbal && NULL != angle1) {
-            gimbal->z_gimbal += angle;
-            *angle1 = gimbal->z_gimbal;
-        }
-
-    } else if (get_vec_gimbal(axis) == GIMBAL_Y) {
-        // rotate around y-axis
-        // pre-processing automata
-        // Let the X-axis in the R(3) space be the X-axis in the 2D projection.
-        // Let the Z-axis in the R(3) space be the Y-axis in the 2D projection.
-
-        // work in XZ plane
-        __rotator->element[0][0] = vector2d_cos(angle);
-        __rotator->element[0][2] = -vector2d_sin(angle);
-        __rotator->element[2][0] = vector2d_sin(angle);
-        __rotator->element[2][2] = vector2d_cos(angle);
-        __rotator->element[1][1] = 1;
-
-        if (NULL != gimbal && NULL != angle1) {
-            gimbal->y_gimbal += angle;
-            *angle1 = gimbal->y_gimbal;
-        }
-
-    } else if (get_vec_gimbal(axis) == GIMBAL_X) {
-
-        // rotate around x-axis
-        // pre-processing automata
-        // Let the Z-axis in the R(3) space be the X-axis in the 2D projection.
-        // Let the Y-axis in the R(3) space be the Y-axis in the 2D projection.
-
-        // work in ZY plane
-        __rotator->element[2][2] = vector2d_cos(angle);
-        __rotator->element[2][1] = -vector2d_sin(angle);
-        __rotator->element[1][2] = vector2d_sin(angle);
-        __rotator->element[1][1] = vector2d_cos(angle);
-        __rotator->element[0][0] = 1;
-
-        if (NULL != gimbal && NULL != angle1) {
-            gimbal->x_gimbal += angle;
-            *angle1 = gimbal->x_gimbal;
-        }
-
-    } else {
-        return EINCOMPATTYPE;
-    }
-
-    return PASS;
-}
 
 static inline status_code rotate_about_gimbal(matrix *__rotator,
                                                matrix *__col_vec,
@@ -147,72 +24,6 @@ static inline status_code rotate_about_gimbal(matrix *__rotator,
     out->x = _col_vec->element[0][0];
     out->y = _col_vec->element[1][0];
     out->z = _col_vec->element[2][0];
-
-    return PASS;
-}
-
-static inline status_code rotate_gimbal(vector3d axis, vec_component angle1,
-                                        matrix *__rotator,
-                                        vec3d_gimbal *in_gimbal,
-                                        vec3d_gimbal *out_gimbal,
-                                        vec3d_processors *procs) {
-
-    status_code __code = init_rotator_gimbal(axis, __rotator, angle1,
-                                             NULL, NULL);
-    if (PASS != __code) {
-        if (NULL != procs && NULL != procs->on_op_failed) {
-            procs->on_op_failed(&vec3d_rotate, __code);
-        }
-        return __code;
-    }
-
-    // init the output orientation matrix
-    vec_component _orient_x[3] = {0, 0, 0,};
-    vec_component _orient_y[3] = {0, 0, 0,};
-    vec_component _orient_z[3] = {0, 0, 0,};
-    vec_component *_orient_comps[3] = {_orient_x, _orient_y, _orient_z};
-    matrix __orient = {
-            .element = _orient_comps,
-            .m = 3,
-            .n = 3
-    };
-
-    mat_processors mat_procs = {
-    };
-
-    __code = mat_product(*__rotator,
-                        *(in_gimbal->orientation),
-                        &__orient, mat_procs);
-    if (PASS != __code) {
-        if (NULL != procs && NULL != procs->on_op_failed) {
-            procs->on_op_failed(&vec3d_rotate, __code);
-        }
-        return __code;
-    }
-
-    struct rotation_metadata metadata = {
-            .in_orientation = &__orient,
-            .out_orientation = out_gimbal->orientation
-    };
-
-    mat_procs.on_entry_iterated = &__on_entry_iterated;
-    mat_procs.metadata = &metadata;
-
-    __code = mat_iterate_elements(__orient, ROW_CONVENTION_ITERATOR, mat_procs);
-    if (PASS != __code) {
-        if (NULL != procs && NULL != procs->on_op_failed) {
-            procs->on_op_failed(&vec3d_rotate, __code);
-        }
-        return __code;
-    }
-
-    if (get_vec_gimbal(axis) == GIMBAL_Z) {
-        out_gimbal->z_gimbal = 0;
-    } else if (get_vec_gimbal(axis) == GIMBAL_Y) {
-        out_gimbal->y_gimbal = 0;
-    } else {
-        out_gimbal->x_gimbal = 0;
-    }
 
     return PASS;
 }
